+++ dev/e2e_app/integration_test/permissions/permissions_location_test.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io' as io;
 
+import 'package:e2e_app/keys.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 import '../common.dart';
@@ -58,7 +59,7 @@ void main() {
 
     if (!await Permission.location.isGranted) {
       expect($('Permission not granted'), findsOneWidget);
-      await $('Grant permission').tap();
+      await $(K.grantLocationPermissionButton).tap();
       if (await $.native.isPermissionDialogVisible(timeout: _timeout)) {
         await $.native.selectCoarseLocation();
         await $.native.selectFineLocation();
@@ -91,7 +92,7 @@ void main() {
 
     if (!await Permission.location.isGranted) {
       expect($('Permission not granted'), findsOneWidget);
-      await $('Grant permission').tap();
+      await $(K.grantLocationPermissionButton).tap();
       if (await $.native2.isPermissionDialogVisible(timeout: _timeout)) {
         await $.native2.selectCoarseLocation();
         await $.native2.selectFineLocation();
@@ -114,5 +115,5 @@ void main() {
       await $(RegExp('lng')).waitUntilVisible(timeout: Duration(seconds: 30)),
       findsOneWidget,
     );
-  });
+  }, tags: ['locale_testing_ios']);
 }