+++ dev/e2e_app/integration_test/service_bluetooth_test.dart
@@ -1,12 +1,19 @@
+import 'dart:io';
+
 import 'common.dart';
 
 void main() {
   patrol('disables and enables bluetooth twice', ($) async {
     await createApp($);
-
-    await $.native.disableBluetooth();
-    await $.native.enableBluetooth();
-    await $.native.disableBluetooth();
-    await $.native.enableBluetooth();
+    // TODO: will be handled with tags in the future
+    if (await $.native.isVirtualDevice() && Platform.isIOS) {
+      $.log('Test will be skipped because of iOS simulator limitations');
+      return;
+    } else {
+      await $.native.disableBluetooth();
+      await $.native.enableBluetooth();
+      await $.native.disableBluetooth();
+      await $.native.enableBluetooth();
+    }
   });
 }