+++ dev/e2e_app/lib/notifications_screen.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io' as io;
 
+import 'package:e2e_app/keys.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:timezone/timezone.dart' as tz;
@@ -135,11 +136,13 @@ class _NotificationsScreenState extends State<NotificationsScreen> {
               spacing: 8,
               children: [
                 ElevatedButton(
+                  key: K.showNotificationNowButton,
                   onPressed: () async =>
                       _showNotificationNow(id: 1, title: _firstTitle),
                   child: const Text('Show now'),
                 ),
                 ElevatedButton(
+                  key: K.showNotificationLaterButton,
                   onPressed: () async =>
                       _showNotificationLater(id: 1, title: _firstTitle),
                   child: const Text('Show in a few seconds'),
@@ -156,11 +159,13 @@ class _NotificationsScreenState extends State<NotificationsScreen> {
               spacing: 8,
               children: [
                 ElevatedButton(
+                  key: K.showNotificationNowButton2,
                   onPressed: () async =>
                       _showNotificationNow(id: 2, title: _secondTitle),
                   child: const Text('Show now'),
                 ),
                 ElevatedButton(
+                  key: K.showNotificationLaterButton2,
                   onPressed: () async =>
                       _showNotificationLater(id: 2, title: _secondTitle),
                   child: const Text('Show in a few seconds'),